refactor: store root runtime component and runtime page in services
