+++ libs/frontend/application/element/src/services/clone-element.service.ts
@@ -269,7 +269,6 @@ export class CloneElementService
       closestContainerNode: element.closestContainerNode,
       id: v4(),
       name: duplicateName,
-      nextSibling: element.nextSibling?.current,
       page: element.page ? { id: element.page.id } : null,
       parentComponent: element.parentComponent
         ? { id: element.parentComponent.id }
