+++ libs/frontend/application/renderer/src/model/runtime-page.model.ts
@@ -7,22 +7,14 @@ import {
   getRuntimeElementService,
   IElementTreeViewDataNode,
   IRuntimeElementModel,
-  runtimePageRef,
-  runtimeStoreRef,
 } from '@codelab/frontend/abstract/application'
-import {
-  type IPageModel,
-  pageRef,
-  storeRef,
-} from '@codelab/frontend/abstract/domain'
+import { type IPageModel } from '@codelab/frontend/abstract/domain'
 import type { Maybe } from '@codelab/shared/abstract/types'
 import { Nullable } from '@codelab/shared/abstract/types'
 import { computed } from 'mobx'
 import type { Ref } from 'mobx-keystone'
 import { idProp, Model, model, prop } from 'mobx-keystone'
 import type { ReactElement } from 'react'
-import { v4 } from 'uuid'
-import { RuntimeStoreModel } from './runtime-store.model'
 
 /**
  * Create both provider page and regular page recursively
@@ -36,35 +28,8 @@ const compositeKey = (page: IPageModel) => `runtime.${page.id}`
 const compositeKeyForProvider = (page: IPageModel, provider: IPageModel) =>
   `runtime.${page.id}.${provider.id}`
 
-const create = (page: IPageModel): IRuntimePageModel => {
-  const runtimeStore = RuntimeStoreModel.create({
-    id: v4(),
-    store: storeRef(
-      page.providerPage ? page.providerPage.store.id : page.store.id,
-    ),
-  })
-
-  const childPage = page.providerPage
-    ? new RuntimePageModel({
-        compositeKey: compositeKey(page),
-        page: pageRef(page.id),
-        runtimeStore: RuntimeStoreModel.create({
-          id: v4(),
-          runtimeProviderStore: runtimeStoreRef(runtimeStore.id),
-          store: storeRef(page.store.id),
-        }),
-      })
-    : undefined
-
-  return new RuntimePageModel({
-    childPage: childPage ? runtimePageRef(childPage) : undefined,
-    compositeKey: page.providerPage
-      ? compositeKeyForProvider(page, page.providerPage)
-      : compositeKey(page),
-    page: pageRef(page.providerPage ? page.providerPage : page),
-    runtimeStore,
-  })
-}
+const create = (dto: IRuntimePageDTO): IRuntimePageModel =>
+  new RuntimePageModel(dto)
 
 @model('@codelab/RuntimePage')
 export class RuntimePageModel
