+++ libs/frontend/application/renderer/src/renderer.model.ts
@@ -25,7 +25,6 @@ import {
 import { computed } from 'mobx'
 import type { ObjectMap, Ref } from 'mobx-keystone'
 import { idProp, Model, model, prop } from 'mobx-keystone'
-import { RuntimeComponentModel, RuntimePageModel } from './model'
 import { defaultPipes, renderPipeFactory } from './render-pipes'
 import { ExpressionTransformer } from './services'
 import { typedPropTransformersFactory } from './typed-prop-transformers'
